refactor: 숫자로 이루어져 있는 지 검증
- 함수 이름을 바꾸어 생성자는 바깥으로 추출
